Remove Korean comments from MSharedCommandTable.h, add English descriptions of ranges, remove XTrap and HShield commands and handlers, replace inet_ntoa to stop warnings, some other stuff I forgot

+++ RealSpace2/Include/RBspObjectDrawD3D9.cpp
@@ -6,7 +6,7 @@
 
 using namespace rsx;
 using namespace RealSpace2;
-using IndexType = u32;
+using IndexType = u16;
 
 template <typename VertexType, typename IndexType>
 static bool CreateBuffers(size_t VertexCount, size_t IndexCount, u32 FVF,
